Simple foraging environment
First commit of the foraging environment. Currently the ForagingEnvNeuromod.py file contains code to construct a simple single-agent environment with one "safehouse" and a "bear". The agent gets reward for reaching the safehouse and a penalty for venturing into the bear territory. Currently, the agent can move in: up, down, left, right manner. The bear is static.